pdf-extractor-api/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
openai
python-dotenv
